cram/record: Wrap reference ID

+++ noodles-cram/src/container/slice.rs
@@ -99,7 +99,7 @@ impl Slice {
                 mate.read_name.extend(record.read_name.iter());
             }
 
-            record.next_fragment_reference_sequence_id = mate.reference_id;
+            record.next_fragment_reference_sequence_id = i32::from(mate.reference_id);
             record.next_mate_alignment_start = mate.alignment_start;
         }
